fixed the office hour model
